% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_reformatting.R
\name{fmtscientific}
\alias{fmtscientific}
\title{Formatters}
\usage{
fmtscientific(number)
}
\arguments{
\item{number}{number to convert into scientific notation}
}
\value{
scientific notation form(character)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\examples{
fmtscientific(0.1)
}
