% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_text.R
\name{fmtsuccess}
\alias{fmtsuccess}
\title{Colour text}
\usage{
fmtsuccess(...)
}
\arguments{
\item{...}{(string/s) Text to colorise. Comma separated strings will be concatenated (no spaces) before colorisation.}
}
\value{
(string) Input text flanked by relevant Ansi escape codes
}
\description{
A collection of functions that take text and return that same text flanked by  characters
that will lead to its coloration/formatting when printed to terminals using message/cat. Different presets are available:
fmterror, fmtwarning, fmtsuccess, fmtbold.

If greater control is required, use the crayon package
}
