% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicode.R
\name{unicode_preview}
\alias{unicode_preview}
\title{List supported unicode paramateres}
\usage{
unicode_preview()
}
\value{
run for its side effects
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\examples{
unicode_preview()
}
