% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilitybeltfmt-package.R
\docType{package}
\name{utilitybeltfmt-package}
\alias{utilitybeltfmt}
\alias{utilitybeltfmt-package}
\title{utilitybeltfmt: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/selkamand/utilitybeltfmt}
  \item Report bugs at \url{https://github.com/selkamand/utilitybeltfmt/issues}
}

}
\author{
\strong{Maintainer}: Sam El-Kamand \email{selkamand@ccia.org.au} (\href{https://orcid.org/0000-0003-2270-8088}{ORCID})

}
\keyword{internal}
