% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_reformatting.R
\name{fmtscientific}
\alias{fmtscientific}
\title{Formatters}
\usage{
fmtscientific(number)
}
\arguments{
\item{number}{number to convert into scientific notation}
}
\value{
scientific notation form(character)
}
\description{
Formatters
}
\examples{
fmtscientific(0.1)
}
