% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_convenience_wrappers.R
\name{message_bold}
\alias{message_bold}
\title{Coloured Message}
\usage{
message_bold(...)
}
\arguments{
\item{...}{(string/s) MEssage to format Comma separated strings will be concatenated (no spaces) before colorisation.}
}
\value{
run for its side effects
}
\description{
Coloured Message
}
