% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_convenience_wrappers.R
\name{message_bold}
\alias{message_bold}
\title{Coloured Message}
\usage{
message_bold(..., symbol = TRUE)
}
\arguments{
\item{...}{(string/s) MEssage to format Comma separated strings will be concatenated (no spaces) before colorisation.}

\item{symbol}{add a symbol indicating the type of message as prefix (boolean) (TRUE)}
}
\value{
run for its side effects
}
\description{
Coloured Message
}
