% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_convenience_wrappers.R
\name{message_error}
\alias{message_error}
\title{Coloured Message}
\usage{
message_error(..., symbol = TRUE, level = 1, verbose = TRUE)
}
\arguments{
\item{...}{(string/s) Message to format. Comma separated strings will be concatenated (no spaces) before colorisation.}

\item{symbol}{add a symbol indicating the type of message as prefix (boolean)}

\item{level}{how many tabs should be used as a prefix to the text (calculated as level-1) (numeric)}

\item{verbose}{should the message be printed? This argument allows verbosity arguments of parent functions to easily inform what messages end-users will see (boolean)}
}
\value{
run for its side effects
}
\description{
Coloured Message
}
\examples{
message_error("This is an error")
message_warning("This is a warning")
message_info("This text is informative")
message_bold("This text is bold")
}
