% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicode.R
\docType{data}
\name{unicode_symbols}
\alias{unicode_symbols}
\title{Unicode Character Codes}
\format{
a list. Names = descriptors of the unicode symbol. Values = unicode code
}
\usage{
unicode_symbols
}
\description{
Unicode character codes
}
\keyword{datasets}
